bash shell options
I took the time to read the bash manuals and turns out you can pretty
much do anything you can with ZSH in pure bash itself. Pretty nifty,
huh!

So I went through the GNU help pages and added the options that deemed
fit. They were pretty daunting and this will of course be a work in
progress for the better part of my tweaking future. But these look
reasonable to me so far!