visual appeal defaults for markdown and git commits
I have taken a fancy to the new Goyo plugin developed. I find
distraction free writing to be a plus, and have enabled autocmds to run
Goyo whenever I edit or create a new markdown or a git commit message.

In fact, I am writing the commit message inside of a vim instance that
has Goyo toggled on and it is beautiful!